requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

requests
httpx

redis
celery

sqlalchemy
psycopg2-binary

ffmpeg-python

pillow
moviepy

elevenlabs

playwright

pydub
pyttsx3
pytz